move RawTester test to normal treadle based tests
